the playground gets a controller, a connector, and an engine button
The page was doing five jobs at once: driving Web MIDI, parsing bytes, owning
the player lifecycle, tracking held voices, and rendering. It now does two, and
they are the two only it can do — deciding what a note sounds like, and wiring
the other pieces together.

RealtimePlaybackCtrl owns engine startup, the durable playback, and forwarding.
It is a pipe rather than a manager: an id goes in with a voice, the same id
stops it. MidiConnector owns everything per-note — devices, hot-plug, parsing,
liveId assignment, the (channel, note) map, and every path that produces an
off: key release, retrigger, panic CC, unplug. It emits a sealed event type, so
the page's handler is an exhaustive when and a new event variant cannot be
forgotten.

Held notes now live in one place instead of two, which incidentally fixes the
display collapsing two channels playing the same note.

Starting the engine is a button. Browsers only allow an AudioContext to start
from a user gesture, so a page that boots on mount comes up silently suspended;
"Start Engine" is that gesture, and the state behind it is honest about whether
the engine is down, coming up, or broken.

Which needed the player to be able to say it broke: Player.ensure() used to
leave its deferred uncompleted on failure, so status sat at LOADING forever and
every awaiter hung. It now fails the deferred, reports FAILED, and drops the
memo so the next call genuinely retries. The Play buttons on the song page and
the docs examples show an exclamation and offer that retry.

+++ klang/src/jsMain/kotlin/KlangBenchmark.kt
@@ -190,7 +190,7 @@ class KlangBenchmark(
 
         while (currentVoices < maxVoicesCap) {
             // Calculate current time in seconds for scheduling
-            val currentTimeSec = cursorFrame.toDouble() / sampleRate
+            val currentTimeSec = cursorFrame / sampleRate
 
             // Clear scheduled voices and add ALL voices starting at current time
             scheduler.clearScheduled("benchmark")
